requirements.txt
pyserial==3.5
python-dotenv~=1.0.0
colorama~=0.4.6